Rename jang.11tydata.js to lang.js
