+++ src/site/content/en/en.11tydata.js
@@ -1,5 +1,6 @@
 const livePaths = require('../../_filters/live-paths');
 const allPaths = require('../../_data/paths');
+const lang = require('./lang');
 
 // =============================================================================
 // HOME OVERVIEW
@@ -19,10 +20,9 @@ module.exports = function () {
     allPaths['secure'],
   ].filter(livePaths);
 
-  const lang = 'en';
-
   return {
-    lang,
+    lang: lang.lang,
+    locale: lang.locale,
     home: {
       paths,
     },
